Update run.sh script to actually compile only the executable

+++ run.sh
@@ -1,19 +1,19 @@
-#!/bin/sh
+#!/bin/bash
 
-gradle build --console=rich
+TYPE="$1"
+if [ -z "$TYPE" ]; then
+    TYPE="debug"
+fi
+TYPE=${TYPE,,}
 
-if [ $? -eq 0 ]; then
-    TYPE="$1"
-    if [ -z "$TYPE" ]; then
-        TYPE="debug"
-    fi
-    rm error.log
+test -f error.log && rm error.log
 
-    printf "\nRunning $TYPE executable...\n"
+if gradle installRtexplore${TYPE^}Executable --console=rich; then
+
+    printf "\nRunning %s executable...\n" $TYPE
     sleep 0.5
 
-    #LD_LIBRARY_PATH=build/libs/rayterm/shared/$TYPE/ ./build/exe/rtexplore/$TYPE/rtexplore 2> error.log
-    xterm -T "i3-float" -geometry 100x50+200+150 -fa "AnonymousProMinus Nerd Font" -fs 8 -e ./build/exe/rtexplore/$TYPE/rtexplore 2> error.log
+    xterm -T "i3-float" -geometry 100x50+200+150 -fa "AnonymousProMinus" -fs 8 -e ./build/install/rtexplore/$TYPE/rtexplore 2> error.log
 
     LOG=$(cat error.log)
